Move storm methods to view
